45-hr-analytics-dashboard/requirements.txt
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.14.1
dash==2.11.1
jupyter==1.0.0
openpyxl==3.1.2
xlrd==2.0.1
joblib==1.2.0
flask==2.3.2
python-dotenv==1.0.0